// src/app/core/backend/elites/models/elite-vm.ts
/* tslint:disable */
/* eslint-disable */
export interface EliteVm {
  attachments?: Array<string> | null;
  bestAchievement?: string | null;
  category?: string | null;
  categoryId?: string | null;
  city?: string | null;
  cityId?: string;
  email?: string | null;
  id?: string;
  mainCategory?: string | null;
  mainCategoryId?: string | null;
  name?: string | null;
  notes?: string | null;
  phoneNumber?: string | null;
  position?: string | null;
  profilePictures?: Array<string> | null;
  qualification?: string | null;
  qualificationId?: string | null;
  socialUrl?: string | null;
  workplace?: string | null;
}